fix: resolve React fast refresh warning in CommunityPrivileges.tsx
- Move CommunityPrivileges_Compact to separate file (CommunityPrivilegesCompact.tsx)
- Rename export to CommunityPrivilegesCompact for consistency
- Ensures proper React Fast Refresh for both components
